Move _infer_date_format to _table_vectorizer
